requirements.txt
clickhouse_connect==0.8.15
pandas==2.2.3
prettytable==3.15.1
python-dotenv==1.0.1
